First claude attempt with implementing cloud_microphysics_2m into icon_physics.
1. main cloud_microphysics_2m function created, mimicking the physics of cloud_micro_inferface in ECHAM
2. apply_microphysics_2m created in icon_physics
3. droplet number is prescribed from plume scheme. ice characteristic radius and number also need prescibed (calculated from cirrus scheme in ECHAM)

+++ jcm/physics/icon/clouds/cloud_params_2m.py
@@ -221,7 +221,7 @@ class CloudParams2M: #(NamedTuple):
     ) -> 'CloudParams2M':
         """Return default cloud parameters for 2-m scheme"""
         # derived helpers
-        disp = float(jnp.exp(0.5 * (sigmaw ** 2)))
+        disp = jnp.exp(0.5 * (sigmaw ** 2))
         dw0 = 10e-6 * disp
         xsec = 1.0 - epsec
         qsec = 1.0 - cqtmin
